Add roles to the params set
Introduce functions for each parameter. This change adds an option
to control what tier/role a parameter is intended for. With this
change the parameters will directed to the right place.

This initial implementation starts with client and server roles.
The keyword 'role' is optional and defaults to 'client' if omitted.

Closes #11