Fix 'enabled' key after 1st include
We delete the 'enabled' key after we consume it to not add more
noise to the input json files.

The 'enabled' key was being removed after 1st include so all other
includes ignored it and enabled it even if enabled value was "no".

This patch fixes the enable key deletion after it is consumed first
time. Instead of deleting the original object, we delete the copy.

Closes #73